delay for 500ms while the storage emulator starts up,
and switch to using "AzureStorageIntegrationTests" collection so that we only have to wait for it to start up once.
