+++ test/Serilog.Sinks.AzureTableStorageWithProperties.Tests/Infrastructure/AzureStorageEmulatorCollection.cs
@@ -1,0 +1,9 @@
+﻿using Xunit;
+
+namespace Serilog.Sinks.AzureTableStorage.Tests.Infrastructure
+{
+    [CollectionDefinition("AzureStorageIntegrationTests")]
+    public class AzureStorageEmulatorCollection : ICollectionFixture<AzureStorageEmulatorFixture>
+    {
+    }
+}
